requirements.txt
flask==3.1.0
pytest==8.3.5
requests==2.32.3
pyyaml==6.0.2
